feat: add useHover composable for hover state management in Carousel
