+++ src/composables/useHover.ts
@@ -1,0 +1,19 @@
+import { ref } from 'vue'
+
+export function useHover() {
+  const isHover = ref(false)
+
+  const handleMouseEnter = (): void => {
+    isHover.value = true
+  }
+
+  const handleMouseLeave = (): void => {
+    isHover.value = false
+  }
+
+  return {
+    isHover,
+    handleMouseEnter,
+    handleMouseLeave,
+  }
+}
